+++ core/src/main/kotlin/com/huanshankeji/exposedvertxsqlclient/EvscInternalApi.kt
@@ -10,4 +10,9 @@ package com.huanshankeji.exposedvertxsqlclient
  */
 @RequiresOptIn("This API is internal in the Exposed Vert.x SQL Client library. It may be changed in the future without notice.")
 @Retention(AnnotationRetention.BINARY)
+annotation class EvscInternalApi
+
+@Deprecated("Use `EvscInternalApi` instead.", ReplaceWith("EvscInternalApi"))
+@RequiresOptIn("This API is internal in the Exposed Vert.x SQL Client library. It may be changed in the future without notice.")
+@Retention(AnnotationRetention.BINARY)
 annotation class InternalApi