+++ integrated/src/benchmarks/kotlin/com/huanshankeji/exposed/benchmark/jdbc/TransactionBenchmark.kt
@@ -1,5 +1,9 @@
-package com.huanshankeji.exposed.benchmark
+package com.huanshankeji.exposed.benchmark.jdbc
 
+import com.huanshankeji.exposed.benchmark.`10K`
+import com.huanshankeji.exposed.benchmark.awaitAsync10K
+import com.huanshankeji.exposed.benchmark.multiThread_10K_ops_nearlyEvenlyPartitioned_helper
+import com.huanshankeji.exposed.benchmark.numProcessors
 import com.huanshankeji.kotlinx.coroutines.benchmark.ParameterizedRunBlockingAwaitAsyncsBenchmark
 import com.huanshankeji.kotlinx.coroutines.benchmark.RunBlockingAwaitAsyncsBenchmark
 import kotlinx.benchmark.*
@@ -41,7 +45,7 @@ class TransactionBenchmark : WithContainerizedDatabaseAndExposedDatabaseBenchmar
 
     @Setup
     fun printNumProcessors() {
-        println("Number of processors: $numProcessors")
+        println("Number of processors: ${numProcessors}")
     }
 
     @Benchmark